+++ server/app.js
@@ -5,6 +5,7 @@ const dotenv = require('dotenv');
 const passportConfig = require('./config/passport');
 const uploadRoute = require("./routes/upload");
 
+
 // Load environment variables
 dotenv.config();
