+++ server/routes/upload.js
@@ -2,19 +2,8 @@ const express = require("express");
 const multer = require("multer");
 const path = require("path");
 const router = express.Router();
+const upload = require("../middleware/upload"); 
 
-// Store uploads in /uploads
-const storage = multer.diskStorage({
-  destination: (req, file, cb) => {
-    cb(null, "uploads/");
-  },
-  filename: (req, file, cb) => {
-    const ext = path.extname(file.originalname);
-    cb(null, `${Date.now()}${ext}`);
-  },
-});
-
-const upload = multer({ storage });
 
 router.post("/", upload.single("image"), (req, res) => {
   if (!req.file) return res.status(400).json({ error: "No file uploaded" });
